--- app/README.md
@@ -1,38 +1,0 @@
-# Aplication SDC
-
-## Objetivo
-
-La aplicacion tiene como objetivo un envio seguro de credenciales a traves de un cifrado asimetrico.
-
-### Funcionalidades:
-
-- Generacion de claves pares
-- Encriptacion de texto
-- Des encriptacion de texto
-- Generacion de password random
-
-## Enviroment
-
-**Windows**
-
-Python 3.9
-
-**MacOS**
-
-Python 3.10+
-
-En algunas versiones de Python en MacOs no viene nativo Tkinter por lo que hay que instalarlo
-
-`brew install python-tk`
-
-
-### Requirements
-
-- Cryptography 39.0.0
-- Tkinter (Nativo)
-- Pathlib (Nativo)
-- Pyperclip 1.8.2
-- PyInstaller 5.7.0
-- (MacOS) Pillow 9.4.0
-
-
