--- app/requirements.txt
@@ -1,5 +1,0 @@
-cryptography==39.0.0
-pyperclip==1.8.2
-pyinstaller==5.7.0
-##for MacOS
-Pillow==9.4.0
