+++ pyproject.toml
@@ -1,0 +1,18 @@
+[tool.poetry]
+name = "python-crypto"
+version = "0.1.0"
+description = "CLI Mensajeria encriptada"
+authors = ["Kevin Barroso <kevin.barroso@taligent.com.ar>"]
+readme = "README.md"
+packages = [{include = "python_crypto"}]
+
+[tool.poetry.dependencies]
+python = "^3.9"
+cryptography = "39"
+pyperclip = "1.8.2"
+typer = {extras = ["all"], version = "^0.7.0"}
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
